Add basic dropdown example
Moodle doesn't have a core template setup to produce a dropdown so
this commit shows what's involved in producing a Moodle compatible
dropdown.

There is an odd side effect in the UI when the button is the sole
element on the page, inside the main content region. The dropdown
element fails to overflow the region, capping it's appearance. This
of course is unlikely to be an issue in production since there is
going to be more than a single button in the content. I have
resisted the temptation to add custom CSS but adding a min height
to the aria-label Content selector would fix this.